Created a distinct mode if the user wants to see live data visualised that syncs data fetch, process, and visualisation.
There appears to be a bug that causes visualised data (SwyftX website) to be desynchronised from the server, which could be corrected by refreshing the page. It didn't affect the way in which we use the API, but it has wasted so much of my time trying to figure out why my data is wrong.

Still need to figure out a way to display multiple graphs at once.